requirements.txt
django
asgiref==3.8.1
djangorestframework==3.15.2
gunicorn
psycopg2-binary==2.9.9
